requirements.txt
flask
flask-cors
mediapipe
opencv-python
numpy
pillow